+++ src/validators/string/same-as.ts
@@ -1,6 +1,7 @@
 // src/validators/string/same-as.ts
 
-import { ValidationRule } from "../../types";
+import { isTypeOfExpression } from "typescript";
+import { FieldConfig, ValidationRule } from "../../types";
 
 /**
  * SameAs - Compares two values for equality
@@ -13,14 +14,17 @@ import { ValidationRule } from "../../types";
  * @returns Validation function that returns error message or empty string
  */
 export const sameAs = (
-    compareValue: (() => any) | any,
+    compareValue: (() => any) | FieldConfig | any,
     message?: string
 ): ValidationRule => {
     return (value: any): string | null => {
         if (!value) return null;
-
-        const valueToCompare =
+        // If i'ts a function execute to get the actual value else return the type as is
+        let valueToCompare =
             typeof compareValue === "function" ? compareValue() : compareValue;
+
+        // If its of type object and can resolve to FieldConfig extract only the value else return as is
+        valueToCompare = (typeof valueToCompare === 'object') ? (valueToCompare as FieldConfig).value : valueToCompare;
 
         if (value !== valueToCompare) {
             return message || "Values don't match";